Create pipfile for pipenv, fix small error

+++ Map.py
@@ -22,8 +22,8 @@ class Map(object):
 
     def render_ontop_of_player(self, screen):
         layer = 5
-        for x in range(0, screen.get_width() / 32):
-            for y in range(0, screen.get_height() / 32):
+        for x in range(0, int(screen.get_width() / 32)):
+            for y in range(0, int(screen.get_height() / 32)):
                 tx = x + self.camera_x
                 ty = y + self.camera_y
 
@@ -43,8 +43,8 @@ class Map(object):
             self.counter = 0
 
         for layer in [0, 1, 4, 5, 7]: # ground, fringe, water, trees, logs
-            for x in range(0, screen.get_width() / 32):
-                for y in range(0, screen.get_height() / 32):
+            for x in range(0, int(screen.get_width() / 32)):
+                for y in range(0, int(screen.get_height() / 32)):
                     tx = x + self.camera_x
                     ty = y + self.camera_y
